refactor(shared): relocate the Databricks SDK facade to shared
Move the workspace-client facade from packages/appkit/src/workspace-client
to packages/shared/src/workspace-client so both appkit and the CLI (which
lives in shared) reach the SDK through one sanctioned import site.

- appkit's workspace-client becomes a thin re-export from
  shared/workspace-client; its ~47 importers are unchanged.
- biome noRestrictedImports allow-list repoints to the new location.
- shared gains the @databricks/sdk-experimental dep and a ./workspace-client
  export (tsdown entry added so it's emitted independently); appkit drops the
  now-unused direct SDK dep.
- Adds a `profile` option to WorkspaceClientOptions.

Pure relocation — no behavior change beyond the additive `profile` option.
Extracted from the registry-CLI work so it can land on its own.
