+++ packages/shared/package.json
@@ -15,6 +15,10 @@
       "development": "./src/cli/index.ts",
       "default": "./dist/cli/index.js"
     },
+    "./workspace-client": {
+      "development": "./src/workspace-client/index.ts",
+      "default": "./dist/workspace-client/index.js"
+    },
     "./package.json": "./package.json"
   },
   "scripts": {
@@ -36,13 +40,15 @@
     "exports": {
       ".": "./dist/index.js",
       "./cli": "./dist/cli/index.js",
+      "./workspace-client": "./dist/workspace-client/index.js",
       "./package.json": "./package.json"
     }
   },
   "dependencies": {
     "@ast-grep/napi": "0.37.0",
-    "@standard-schema/spec": "1.1.0",
     "@clack/prompts": "1.0.1",
+    "@databricks/sdk-experimental": "0.17.0",
+    "@standard-schema/spec": "1.1.0",
     "commander": "12.1.0",
     "dotenv": "16.6.1",
     "js-yaml": "4.2.0",
