+++ packages/shared/src/workspace-client/errors.ts
