+++ packages/shared/src/workspace-client/tests/legacy.test.ts
